+++ Fabric/src/main/java/net/potionstudios/biomeswevegone/fabric/VanillaCompatFabric.java
@@ -20,7 +20,6 @@ import net.potionstudios.biomeswevegone.world.entity.pumpkinwarden.PumpkinWarden
 import net.potionstudios.biomeswevegone.world.item.BWGItems;
 import net.potionstudios.biomeswevegone.world.item.brewing.BWGBrewingRecipes;
 import net.potionstudios.biomeswevegone.world.item.tools.ToolInteractions;
-import net.potionstudios.biomeswevegone.world.level.block.BWGBlocks;
 import net.potionstudios.biomeswevegone.world.level.block.BlockFeatures;
 import net.potionstudios.biomeswevegone.config.configs.BWGWorldGenConfig;
 import net.potionstudios.biomeswevegone.world.level.levelgen.biome.modifiers.BWGBiomeModifiers;
@@ -34,7 +33,7 @@ public class VanillaCompatFabric {
     public static void init() {
         ToolInteractions.registerStrippableBlocks(StrippableBlockRegistry::register);
         BlockFeatures.registerFlammable(FlammableBlockRegistry.getDefaultInstance()::add);
-        registerFuels();
+        BlockFeatures.registerFurnaceFuels((item, burnTime) -> FuelRegistryEvents.BUILD.register(((builder, context) -> builder.add(item, burnTime))));
         BlockFeatures.registerCompostables(CompostingChanceRegistry.INSTANCE::add);
         ToolInteractions.registerFlattenables(FlattenableBlockRegistry::register);
         ToolInteractions.registerTillables((block, pair) -> TillableBlockRegistry.register(block, pair.getFirst(), pair.getSecond()));
@@ -48,10 +47,6 @@ public class VanillaCompatFabric {
         FabricBrewingRecipeRegistryBuilder.BUILD.register(builder -> BWGBrewingRecipes.buildBrewingRecipes(builder::addMix));
         BWGVillagerType.setVillagerBiomes(VillagerType.BY_BIOME::put);
         UseEntityCallback.EVENT.register(((player, level, interactionHand, entity, entityHitResult) -> PumpkinWarden.villagerToPumpkinWarden(player, player.getItemInHand(interactionHand), level) ? InteractionResult.SUCCESS : InteractionResult.PASS));
-    }
-
-    private static void registerFuels() {
-        FuelRegistryEvents.BUILD.register(((builder, context) -> builder.add(BWGBlocks.PEAT.get(), 1200)));
     }
 
     private static void registerBiomeModifiers() {