+++ Forge/src/main/java/net/potionstudios/biomeswevegone/forge/VanillaCompatForge.java
@@ -58,7 +58,6 @@ public class VanillaCompatForge {
         }
         bus.addListener(VanillaCompatForge::registerBrewingRecipes);
         bus.addListener(VanillaCompatForge::onBoneMealUse);
-        //bus.addListener(VanillaCompatForge::onEnderManAnger);
         bus.addListener(VanillaCompatForge::onVillagerInteract);
     }
 
@@ -83,8 +82,9 @@ public class VanillaCompatForge {
      * @see FurnaceFuelBurnTimeEvent
      */
     private static void registerFuels(final FurnaceFuelBurnTimeEvent event) {
-        if (event.getItemStack().is(BWGBlocks.PEAT.get().asItem()))
-            event.setBurnTime(1200);
+        BlockFeatures.registerFurnaceFuels((block, burnTime) -> {
+            if (event.getItemStack().is(block.asItem())) event.setBurnTime(burnTime);
+        });
     }
 
     /**