+++ NeoForge/src/main/java/net/potionstudios/biomeswevegone/neoforge/datagen/generators/DatamapGenerator.java
@@ -14,7 +14,6 @@ import net.neoforged.neoforge.registries.datamaps.builtin.FurnaceFuel;
 import net.neoforged.neoforge.registries.datamaps.builtin.NeoForgeDataMaps;
 import net.potionstudios.biomeswevegone.BiomesWeveGone;
 import net.potionstudios.biomeswevegone.world.entity.npc.BWGVillagerType;
-import net.potionstudios.biomeswevegone.world.level.block.BWGBlocks;
 import net.potionstudios.biomeswevegone.world.level.block.BlockFeatures;
 import org.jetbrains.annotations.NotNull;
 
@@ -33,14 +32,13 @@ public class DatamapGenerator extends DataMapProvider {
 
     @Override
     protected void gather(HolderLookup.@NotNull Provider arg) {
-        builder(NeoForgeDataMaps.FURNACE_FUELS)
-                .add(id(BWGBlocks.PEAT.get().asItem()), new FurnaceFuel(1200), false)
-                .conditions(new ModLoadedCondition(BiomesWeveGone.MOD_ID));
+        Builder<FurnaceFuel, Item> fuelBuilder = builder(NeoForgeDataMaps.FURNACE_FUELS);
+        BlockFeatures.registerFurnaceFuels((block, burnTime) -> fuelBuilder.add(id(block.asItem()), new FurnaceFuel(burnTime), false));
+        fuelBuilder.conditions(new ModLoadedCondition(BiomesWeveGone.MOD_ID));
 
-        var builder = builder(NeoForgeDataMaps.COMPOSTABLES);
-        BlockFeatures.registerCompostables((item, chance) -> builder.add(id(item.asItem()), new Compostable(chance, true), false));
-        builder.conditions(new ModLoadedCondition(BiomesWeveGone.MOD_ID));
-
+        Builder<Compostable, Item> compostableBuilder = builder(NeoForgeDataMaps.COMPOSTABLES);
+        BlockFeatures.registerCompostables((item, chance) -> compostableBuilder.add(id(item.asItem()), new Compostable(chance, true), false));
+        compostableBuilder.conditions(new ModLoadedCondition(BiomesWeveGone.MOD_ID));
 
         Builder<BiomeVillagerType, Biome> biomeVillagerTypeBuilder = builder(NeoForgeDataMaps.VILLAGER_TYPES);
         BWGVillagerType.setVillagerBiomes(((biomeResourceKey, villagerType) -> biomeVillagerTypeBuilder.add(biomeResourceKey, new BiomeVillagerType(villagerType), false)));